+++ docs/Whats-New.md
@@ -136,6 +136,7 @@ New:
 - Add the ability to queue/dequeue 5 units at a time, or dequeue all units instantly (by ZivDero)
 - Make OverlayTypes 27 to 38 (fourth Tiberium images) passable by infantry (by AlexB)
 - Implement the support for new ArmorTypes and allow forbidding force-fire, passive-acquire and retaliation versus specific armor types (by ZivDero/CCHyper)
+- Add a developer command to dump all heaps to the log (by ZivDero)
 
 Vanilla fixes:
 - Fix HouseType `Nod` having the `Prefix=B` and `Side=GDI` in vanilla `rules.ini` by setting them to `N` and `Nod`, respectively (by CCHyper/tomsons26)